Se creó package miJuego. Se importo el package en las clases necesarias. Se creó sistema de combate y se configuró el Menu en base a eso. Se simplificó el Main.

+++ src/Elfo.java
@@ -2,7 +2,7 @@ import miJuego.Personaje;
 
 public class Elfo extends Personaje {
 
-    public Elfo(Raza raza, String nombre, int edad, int salud, int velocidad, int destreza, int fuerza, int armadura, int nivel) {
+    public Elfo(Raza raza, String nombre, int edad, float salud, int velocidad, int destreza, int fuerza, int armadura, int nivel) {
         super(raza, nombre, edad, salud, velocidad, destreza, fuerza, armadura, nivel);
     }
