+++ src/Orco.java
@@ -1,7 +1,7 @@
 import miJuego.Personaje;
 
 public class Orco extends Personaje {
-    public Orco(Raza raza, String nombre, int edad, int salud, int velocidad, int destreza, int fuerza, int armadura, int nivel) {
+    public Orco(Raza raza, String nombre, int edad, float salud, int velocidad, int destreza, int fuerza, int armadura, int nivel) {
         super(raza, nombre, edad, salud, velocidad, destreza, fuerza, armadura, nivel);
     }
